refactor: remove default middleware
Fixed some bugs about handling req/res object on middleware.
