+++ src/middleware.ts
@@ -1,5 +1,1 @@
-import { defaultMiddleware } from './middleware/default'
-
-export class Middleware {
-  static default = defaultMiddleware
-}
+export class Middleware {}
