--- src/middleware/default.test.ts
@@ -1,26 +1,0 @@
-import { Hono } from '../hono'
-
-describe('default middleware', () => {
-  const app = new Hono()
-  app.get('/:foo', (c) => {
-    const query = c.req.query('foo')
-    const param = c.req.param('foo')
-    const header = c.req.header('User-Agent')
-    c.header('X-Query', query)
-    c.header('X-Param', param)
-    c.header('X-Header', header)
-    return c.body('Hono')
-  })
-
-  it('query', async () => {
-    const url = new URL('http://localhost/bar')
-    url.searchParams.append('foo', 'bar')
-    const req = new Request(url.toString())
-    req.headers.append('User-Agent', 'bar')
-    const res = await app.dispatch(req)
-    expect(res.status).toBe(200)
-    expect(res.headers.get('X-Query')).toBe('bar')
-    expect(res.headers.get('X-Param')).toBe('bar')
-    expect(res.headers.get('X-Header')).toBe('bar')
-  })
-})
